FIX RESPONSE > RESULT ADD Service Unit Test

+++ src/main/java/com/jonghae5/jongbirdapi/service/FollowService.java
@@ -2,10 +2,7 @@ package com.jonghae5.jongbirdapi.service;
 
 import com.jonghae5.jongbirdapi.domain.Follow;
 import com.jonghae5.jongbirdapi.domain.User;
-import com.jonghae5.jongbirdapi.exception.post.InvalidateFollowException;
-import com.jonghae5.jongbirdapi.exception.post.InvalidateUnFollowException;
-import com.jonghae5.jongbirdapi.exception.user.InvalidateUserBlockException;
-import com.jonghae5.jongbirdapi.exception.user.InvalidateUserException;
+import com.jonghae5.jongbirdapi.exception.user.*;
 import com.jonghae5.jongbirdapi.repository.follow.FollowRepository;
 import com.jonghae5.jongbirdapi.repository.user.UserRepository;
 import com.jonghae5.jongbirdapi.view.follow.FollowResponse;
@@ -19,6 +16,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @Slf4j
@@ -28,31 +26,49 @@ public class FollowService {
 
     private final UserRepository userRepository;
     private final FollowRepository followRepository;
-    public FollowResponse addFollower(User loginUser, Long userId) {
-        User findUser = userRepository.findById(userId).orElseThrow(InvalidateFollowException::new);
+    public FollowResponse addFollower(User loginUser, Long followUserId) {
+        User findUser = userRepository.findById(followUserId).orElseThrow(InvalidateFollowException::new);
 
+
+        Optional<Follow> exFollowData = followRepository.findByFollowingAndFollower(loginUser, findUser);
+        if (exFollowData.isPresent()) {
+            throw new InvalidateFollowExistException();
+        }
         Follow follow = Follow.builder()
                 .following(loginUser)
                 .follower(findUser)
                 .build();
 
+
         followRepository.save(follow);
 
-        return new FollowResponse(userId);
+        return new FollowResponse(followUserId);
     }
 
-    public FollowResponse deleteFollower(User loginUser, Long userId) {
-        User findUser = userRepository.findById(userId).orElseThrow(InvalidateUnFollowException::new);
+    public FollowResponse deleteFollower(User loginUser, Long followUserId) {
+        User findUser = userRepository.findById(followUserId).orElseThrow(InvalidateUnFollowException::new);
+
+        Optional<Follow> exFollowData = followRepository.findByFollowingAndFollower(loginUser, findUser);
+        if (exFollowData.isEmpty()) {
+            throw new InvalidateFollowUnExistException();
+        }
+
         followRepository.deleteByFollowingAndFollower(loginUser, findUser);
 
-        return new FollowResponse(userId);
+        return new FollowResponse(followUserId);
     }
 
-    public FollowResponse deleteFollowing(User loginUser, Long userId) {
-        User findUser = userRepository.findById(userId).orElseThrow(InvalidateUserBlockException::new);
+    public FollowResponse deleteFollowing(User loginUser, Long followingUserId) {
+        User findUser = userRepository.findById(followingUserId).orElseThrow(InvalidateUserBlockException::new);
+
+        Optional<Follow> exFollowData = followRepository.findByFollowingAndFollower(loginUser, findUser);
+        if (exFollowData.isEmpty()) {
+            throw new InvalidateFollowUnExistException();
+        }
+
         followRepository.deleteByFollowingAndFollower(loginUser, findUser);
 
-        return new FollowResponse(userId);
+        return new FollowResponse(followingUserId);
     }
 
     public GetFollowersResponse getFollowers(User loginUser, int limit) {
